视频上传并制作缩略图功能完成
目前没有图片视频服务器没办法测试。

+++ cilicili-common/src/main/java/com/cilicili/common/utils/VideoResolution.java
@@ -281,7 +281,7 @@ public class VideoResolution {
      * @param picPath 截图存放路径
      * @param iFrame 指定视频帧数
      */
-    public static void getVideoPic(File video, String picPath, int iFrame) {
+    public static void getVideoPic(File video, String picPath, long iFrame) {
         FFmpegFrameGrabber ff = new FFmpegFrameGrabber(video);
         try {
             ff.start();
